+++ src/assets/style/style.scss
@@ -6,8 +6,8 @@
 @import url('https://fonts.googleapis.com/css2?family=Roboto:wght@500;700&display=swap');
 @import url('https://fonts.googleapis.com/css2?family=Roboto:wght@600;700&display=swap');
 
-@import '../../model/todo-list/todo-list.scss';
-@import '../../model/todo-item/todo-item.scss';
+@import '../../view/todo-list/todo-list.scss';
+@import '../../view/todo-item/todo-item.scss';
 
 @import './utils/reset.scss';
 @import './utils/vareable';
