+++ src/view/todo-item/todo-item.scss
